refactor(tests): share analytics market db fixtures
